requirements.txt
Pillow
potracer
